Replace the default versioning method with the new 'AppendedVersioning'
