+++ src/Traits/Versioned.php
@@ -1,8 +1,11 @@
-<?php namespace EloquentVersioned\Traits;
+<?php
+
+namespace EloquentVersioned\Traits;
 
 use EloquentVersioned\Builder as VersionedBuilder;
 use EloquentVersioned\Scopes\VersioningScope;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Model;
 
 trait Versioned
 {
@@ -27,28 +30,6 @@ trait Versioned
     /*
      * ACCESSORS + MUTATORS
      */
-
-    /**
-     * @return mixed
-     */
-    public function getIdAttribute()
-    {
-        return ($this->{static::getIsCurrentVersionColumn()} == 0) ?
-            $this->attributes[$this->primaryKey] :
-            $this->attributes[static::getModelIdColumn()];
-    }
-
-    /*
-     * ELOQUENT OVERRIDES
-     */
-
-    /**
-     * @return string
-     */
-    public function getKeyName()
-    {
-        return $this->isVersioned ? $this->getModelIdColumn() : $this->primaryKey;
-    }
 
     /**
      * @param $query
@@ -105,74 +86,54 @@ trait Versioned
     }
 
     /**
+     * Save a new version of the model
+     *
      * @param array $options
      *
      * @return bool
      */
-    public function save(array $options = array())
+    public function save(array $options = [])
     {
         $query = $this->newQueryWithoutScopes();
 
         $db = $this->getConnection();
 
-        // If the "saving" event returns false we'll bail out of the save and return
-        // false, indicating that the save failed. This provides a chance for any
-        // listeners to cancel save operations if validations fail or whatever.
+        // If the "saving" event returns false we'll bail out of the save and
+        // return false, indicating that the save failed. This provides a chance
+        // for any listeners to cancel save operations if validations fail or
+        // whatever.
         if ($this->fireModelEvent('saving') === false) {
             return false;
         }
 
-        // If the model already exists in the database we can just update our record
-        // that is already in this database using the current IDs in this "where"
-        // clause to only update this model. Otherwise, we'll just insert them.
+        // If the model already exists in the database we can just update our
+        // record that is already in this database using the current IDs in this
+        // "where" clause to only update this model. Otherwise, we'll just
+        // insert them.
         if ($this->exists) {
-            $dirty = $this->getDirty();
+            if ($this->isDirty()) {
+                $saved = $db->transaction(function () use ($query, $db, $options) {
+                    $oldVersion = $this->replicate();
+                    $oldVersion->forceFill($this->original);
+                    $oldVersion->{$this->primaryKey} = null;
+                    $oldVersion->{static::getIsCurrentVersionColumn()} = false;
 
-            if (count($dirty) > 0) {
-                $saved = $db->transaction(function () use ($query, $db) {
-
-                    $newVersion = $this->replicate([
-                        $this->primaryKey,
-                        static::getVersionColumn(),
-                        'updated_at'
-                    ]);
-                    $newVersion->{static::getVersionColumn()} = static::getNextVersion($this->{static::getModelIdColumn()});
-                    $newVersion->{static::getIsCurrentVersionColumn()} = 1;
-                    $newVersion->updated_at = $this->freshTimestamp();
+                    $this->performVersionedInsert($query, $oldVersion);
 
                     // trigger the update event
                     if ($this->fireModelEvent('updating') === false) {
                         return false;
                     }
 
-                    // clear out the old stuff
-                    unset($this->attributes[$this->primaryKey]);
-                    unset($this->attributes['updated_at']);
-
-                    $this->forceFill($newVersion->getAttributes());
-
-                    $saved = $this->performVersionedInsert($query,
-                        ['timestamps' => false]);
+                    $this->{static::getVersionColumn()} = static::getNextVersion($this->{static::getModelIdColumn()});
+                    $saved = $this->performUpdate($query, $options);
 
                     if ($saved) {
-
-                        // toggle the is_current_version flag
-                        $db->table((new static)->getTable())
-                            ->where(static::getModelIdColumn(),
-                                $this->{static::getModelIdColumn()})
-                            ->where(static::getIsCurrentVersionColumn(), 1)
-                            ->where($this->primaryKey, '<>',
-                                $this->attributes[$this->primaryKey])
-                            ->update([static::getIsCurrentVersionColumn() => 0]);
-
                         $this->fireModelEvent('updated', false);
                     }
 
                     return $saved;
-
                 });
-            } else {
-                $saved = true;
             }
         }
 
@@ -202,39 +163,13 @@ trait Versioned
      * EXTENSIONS
      */
 
-    protected function performVersionedInsert(
-        Builder $query,
-        array $options = array()
-    ) {
-        // First we'll need to create a fresh query instance and touch the creation and
-        // update timestamps on this model, which are maintained by us for developer
-        // convenience. After, we will just continue saving these model instances.
-        if ($this->timestamps && array_get($options, 'timestamps', true)) {
-            $this->updateTimestamps();
-        }
-
-        // If the model has an incrementing key, we can use the "insertGetId" method on
-        // the query builder, which will give us back the final inserted ID for this
-        // table from the database. Not all tables have to be incrementing though.
-        $attributes = $this->attributes;
-
-        if ($this->incrementing) {
-            $this->insertAndSetId($query, $attributes);
-        }
-
-        // If the table is not incrementing we'll simply insert this attributes as they
-        // are, as this attributes arrays must contain an "id" column already placed
-        // there by the developer as the manually determined key for these models.
-        else {
-            $query->insert($attributes);
-        }
-
-        // We will go ahead and set the exists property to true, so that it is set when
-        // the created event is fired, just in case the developer tries to update it
-        // during the event. This will allow them to do so and run an update here.
-        $this->exists = true;
-
-        return true;
+    /**
+     * @param Builder $query
+     * @param Model   $model
+     */
+    public function performVersionedInsert(Builder $query, Model $model)
+    {
+        return $query->insert($model->getAttributes());
     }
 
     /**
