--- tests/Models/Thingy.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace EloquentVersioned\Tests\Models;
-
-use EloquentVersioned\Traits\AppendedVersioning;
-use Illuminate\Database\Eloquent\Model;
-
-class Thingy extends Model
-{
-    use AppendedVersioning;
-
-    protected $table = 'thingies';
-
-    protected $fillable = ['*'];
-}
